# contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Jobbely API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/v1/jobs/{id}/requirements": {
      "get": {
        "operationId": "getJobRequirements",
        "parameters": [
          {
            "schema": {
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "version",
                    "contentHash",
                    "category",
                    "skills",
                    "experience",
                    "constraints",
                    "unparsed",
                    "locations",
                    "workplace",
                    "warnings",
                    "truncated"
                  ],
                  "properties": {
                    "version": {
                      "type": "string"
                    },
                    "contentHash": {
                      "type": "string"
                    },
                    "category": {
                      "type": "string"
                    },
                    "skills": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "alternatives",
                          "importance",
                          "evidence"
                        ],
                        "properties": {
                          "alternatives": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "required": [
                                "id",
                                "name"
                              ],
                              "properties": {
                                "id": {
                                  "type": "string"
                                },
                                "name": {
                                  "type": "string"
                                }
                              }
                            }
                          },
                          "importance": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "required"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "preferred"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "contextual"
                                ]
                              }
                            ]
                          },
                          "evidence": {
                            "type": "object",
                            "required": [
                              "excerpt",
                              "line",
                              "rule"
                            ],
                            "properties": {
                              "excerpt": {
                                "type": "string"
                              },
                              "line": {
                                "type": "integer"
                              },
                              "rule": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    },
                    "experience": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "minimumMonths",
                          "scope",
                          "skillId",
                          "importance",
                          "evidence"
                        ],
                        "properties": {
                          "minimumMonths": {
                            "type": "integer"
                          },
                          "scope": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "professional"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "function"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "skill"
                                ]
                              }
                            ]
                          },
                          "skillId": {
                            "type": "string",
                            "nullable": true
                          },
                          "importance": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "required"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "preferred"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "contextual"
                                ]
                              }
                            ]
                          },
                          "evidence": {
                            "type": "object",
                            "required": [
                              "excerpt",
                              "line",
                              "rule"
                            ],
                            "properties": {
                              "excerpt": {
                                "type": "string"
                              },
                              "line": {
                                "type": "integer"
                              },
                              "rule": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    },
                    "constraints": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "kind",
                          "importance",
                          "evidence"
                        ],
                        "properties": {
                          "kind": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "location"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "authorization"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "qualification"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "language"
                                ]
                              }
                            ]
                          },
                          "importance": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "required"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "preferred"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "contextual"
                                ]
                              }
                            ]
                          },
                          "evidence": {
                            "type": "object",
                            "required": [
                              "excerpt",
                              "line",
                              "rule"
                            ],
                            "properties": {
                              "excerpt": {
                                "type": "string"
                              },
                              "line": {
                                "type": "integer"
                              },
                              "rule": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    },
                    "unparsed": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "importance",
                          "evidence"
                        ],
                        "properties": {
                          "importance": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "required"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "preferred"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "contextual"
                                ]
                              }
                            ]
                          },
                          "evidence": {
                            "type": "object",
                            "required": [
                              "excerpt",
                              "line",
                              "rule"
                            ],
                            "properties": {
                              "excerpt": {
                                "type": "string"
                              },
                              "line": {
                                "type": "integer"
                              },
                              "rule": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    },
                    "locations": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "workplace": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "remote"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "hybrid"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "onsite"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "unknown"
                          ]
                        }
                      ]
                    },
                    "warnings": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "truncated": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}